+++ src/main/java/axelmontini/immersivesawmills/common/CommonProxy.java
@@ -35,7 +35,7 @@ public class CommonProxy {
 
     }
 
-    public void spawnParticleOnlyClient(World world, EnumParticleTypes type, float x, float y, float z, float speedX, float speedY, float speedZ, int... params) {
+    public void spawnParticleOnlyClient(World world, EnumParticleTypes type, double x, double y, double z, double speedX, double speedY, double speedZ, int... params) {
 
     }
